autorig-online/mcp/pyproject.toml
[project]
name = "renderfin-mcp"
version = "0.3.0"
description = "MCP server for RenderFin image generation API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "filelock>=3.13.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
renderfin-mcp = "renderfin_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
